test/requirements.txt
dataclasses-json==0.6.1
gitdb==4.0.10
GitPython==3.1.37
iniconfig==2.0.0
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.1
pip==22.3.1
pluggy==1.3.0
pytest==7.4.2
setuptools==65.5.1
smmap==5.0.1
typing-inspect==0.9.0
typing_extensions==4.8.0
wheel==0.38.4
